server/requirements.txt
# Flask core
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
psycopg[binary]==3.1.18
flask-migrate==4.0.5

# Authentication
pyjwt==2.8.0

# WebSocket
python-socketio==5.10.0
eventlet==0.34.2

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Production server
gunicorn==21.2.0

